cephadm: move DaemonIdentity to daemon_identity.py
Pair-programmed-with: Adam King <adking@redhat.com>
